requirements.txt
# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Computer Vision
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0
imgaug>=0.4.0

# ONNX and Optimization
onnx>=1.14.0
onnxruntime>=1.15.0
onnxruntime-gpu>=1.15.0
onnx-simplifier>=0.4.0

# Tracking Libraries
filterpy>=1.4.5
scipy>=1.10.0
scikit-learn>=1.3.0
lap>=0.4.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
PyYAML>=6.0
tqdm>=4.65.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Logging and Monitoring
tensorboard>=2.13.0
wandb>=0.15.0

# Utilities
colorama>=0.4.6
python-dotenv>=1.0.0
rich>=13.0.0
pydantic>=2.0.0

# Video Processing
ffmpeg-python>=0.2.0
moviepy>=1.0.3

# Edge Deployment (Optional - install based on hardware)
# pycuda>=2022.1  # For TensorRT
# onnx-tensorrt>=8.6.0  # For TensorRT conversion

# Development Tools
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
pre-commit>=3.3.0

# Jupyter
jupyter>=1.0.0
ipywidgets>=8.0.0
notebook>=7.0.0
